pyproject.toml
[project]
name = "personal-triage-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=1.0.5",
    "langgraph-checkpoint-sqlite>=1.0.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "pydantic>=2.0.0",
    "chromadb>=0.4.24",
    "sentence-transformers>=2.0.0",
    "pyyaml>=6.0.0",
    "msgpack>=1.0.0",
]
